# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de autenticación con Google",
        "description": "Esta API permite autenticarse con Google y registrar nuevos usuarios.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/disease-patients": {
            "get": {
                "summary": "Obtiene una lista paginada de las relaciones entre enfermedades y pacientes",
                "description": "Obtiene una lista paginada de las relaciones entre enfermedades y pacientes.",
                "operationId": "e2c500c31ee463ece330cf610a08140b",
                "responses": {
                    "200": {
                        "description": "Lista de relaciones entre enfermedades y pacientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/DiseasePatientResource"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string"
                                                },
                                                "last": {
                                                    "type": "string"
                                                },
                                                "prev": {
                                                    "type": "string"
                                                },
                                                "next": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Crea una nueva relación entre enfermedades y pacientes",
                "description": "Crea una nueva relación entre enfermedades y pacientes.",
                "operationId": "f8236cc410b09df3319d55a781e15a5d",
                "requestBody": {
                    "description": "Datos de la relación a crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "patient_user_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "disease_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Relación creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/DiseasePatientResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DiseasePatientResource": {
                "title": "Disease Patient Resource",
                "description": "Disease Patient resource schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "patient_user_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "disease_id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        },
...
    }
}